refactor: modularize monolithic tui.rs into state.rs, draw.rs, and mod.rs submodules
