+++ src/tui/draw.rs
@@ -1,734 +1,20 @@
-use std::io::{self, Stdout};
-use std::path::PathBuf;
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::mpsc::{self, Sender};
-use std::sync::Arc;
-use std::thread;
-use std::time::{Duration, SystemTime};
-
-use ratatui::backend::CrosstermBackend;
-use ratatui::crossterm::event::{self, Event, KeyCode, KeyEventKind};
-use ratatui::crossterm::terminal::{
-    disable_raw_mode, enable_raw_mode, EnterAlternateScreen, LeaveAlternateScreen,
-};
-use ratatui::crossterm::ExecutableCommand;
+use std::time::Duration;
 use ratatui::layout::{Constraint, Direction, Layout, Rect};
 use ratatui::style::{Color, Modifier, Style};
 use ratatui::text::{Line, Span};
 use ratatui::widgets::{Block, BorderType, Borders, Cell, Clear, Paragraph, Row, Table, Tabs};
-use ratatui::Terminal;
 
-use crate::capture::{CaptureLoop, XcapBackend};
-use crate::config::{AppConfig, Theme};
-use crate::doctor::{run_diagnostics, DoctorCheck, CheckStatus};
-use crate::manage::{
-    create_clean_plan, execute_clean_plan, list_sessions, open_session, CleanOptions,
-    SessionSummary, SessionTarget,
-};
-use crate::render::{render, RenderOptions, RenderTarget, parse_exclusions};
-use crate::session::{DisplayTarget, Library, Session, SessionOpenOptions};
+use crate::doctor::CheckStatus;
+use crate::config::Theme;
+use crate::manage::SessionSummary;
+use crate::session::DisplayTarget;
+use crate::render::parse_exclusions;
+use super::state::{ActiveTab, CaptureMode, CaptureState, RenderState, TuiState};
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ActiveTab {
-    Capture = 0,
-    Render = 1,
-    Sessions = 2,
-    Diagnostics = 3,
-}
-
-impl ActiveTab {
-    fn next(self) -> Self {
-        match self {
-            ActiveTab::Capture => ActiveTab::Render,
-            ActiveTab::Render => ActiveTab::Sessions,
-            ActiveTab::Sessions => ActiveTab::Diagnostics,
-            ActiveTab::Diagnostics => ActiveTab::Capture,
-        }
-    }
-
-    fn prev(self) -> Self {
-        match self {
-            ActiveTab::Capture => ActiveTab::Diagnostics,
-            ActiveTab::Render => ActiveTab::Capture,
-            ActiveTab::Sessions => ActiveTab::Render,
-            ActiveTab::Diagnostics => ActiveTab::Sessions,
-        }
-    }
-}
-
-enum TuiMessage {
-    CaptureStarted(Arc<AtomicBool>),
-    FrameCaptured(u64),
-    CaptureFinished(u64),
-    CaptureError(String),
-    RenderProgress(String),
-    RenderFinished(String),
-    RenderError(String),
-}
-
-enum CaptureState {
-    Idle,
-    Starting,
-    Capturing {
-        frames_collected: u64,
-        stop_handle: Arc<AtomicBool>,
-    },
-    Error(String),
-}
-
-enum RenderState {
-    Idle,
-    Rendering(String),
-    Success(String),
-    Error(String),
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum CaptureMode {
-    NewSession,
-    Append,
-}
-
-struct TuiState {
-    active_tab: ActiveTab,
-    library_path: Option<PathBuf>,
-    resolved_library_path: PathBuf,
-    capture_interval: Duration,
-    capture_display: DisplayTarget,
-    capture_state: CaptureState,
-    capture_mode: CaptureMode,
-    render_fps: u32,
-    render_state: RenderState,
-    sessions: Vec<SessionSummary>,
-    cursor_session_index: usize,
-    active_session_index: usize,
-    diagnostics: Option<Vec<DoctorCheck>>,
-    confirm_clean_index: Option<usize>,
-    change_library_input: Option<String>,
-    status_message: Option<(String, SystemTime)>,
-    show_welcome: bool,
-    supports_unicode: bool,
-    config: AppConfig,
-}
-
-fn check_unicode_support() -> bool {
-    // Check standard environment variables
-    for var in &["LANG", "LC_ALL", "LC_CTYPE"] {
-        if let Ok(val) = std::env::var(var) {
-            let val_upper = val.to_uppercase();
-            if val_upper.contains("UTF-8") || val_upper.contains("UTF8") {
-                return true;
-            }
-        }
-    }
-
-    // Windows specific checks
-    if cfg!(target_os = "windows") {
-        // Windows Terminal supports UTF-8 natively
-        if std::env::var("WT_SESSION").is_ok() {
-            return true;
-        }
-        // VS Code terminal, Git Bash, etc.
-        if let Ok(term) = std::env::var("TERM") {
-            if term == "xterm-256color" || term == "cygwin" {
-                return true;
-            }
-        }
-    } else {
-        // macOS and Linux default terminals usually support UTF-8 unless explicitly configured otherwise
-        if std::env::var("TERM").is_ok() {
-            return true;
-        }
-    }
-
-    false
-}
-
-impl TuiState {
-    fn new(library_path: Option<PathBuf>) -> std::result::Result<Self, String> {
-        let config = AppConfig::load();
-        let resolved = library_path
-            .clone()
-            .or_else(|| config.default_library.clone())
-            .or_else(|| Library::default_path().ok())
-            .ok_or_else(|| "Failed to resolve default library path".to_string())?;
-
-        let supports_unicode = check_unicode_support();
-
-        Ok(Self {
-            active_tab: ActiveTab::Capture,
-            library_path,
-            resolved_library_path: resolved,
-            capture_interval: Duration::from_secs(6),
-            capture_display: DisplayTarget::All,
-            capture_state: CaptureState::Idle,
-            capture_mode: CaptureMode::NewSession,
-            render_fps: 15,
-            render_state: RenderState::Idle,
-            sessions: Vec::new(),
-            cursor_session_index: 0,
-            active_session_index: 0,
-            diagnostics: None,
-            confirm_clean_index: None,
-            change_library_input: None,
-            status_message: None,
-            show_welcome: true,
-            supports_unicode,
-            config,
-        })
-    }
-
-    fn refresh_sessions(&mut self) {
-        if let Ok(list) = list_sessions(self.library_path.clone()) {
-            self.sessions = list;
-            if self.cursor_session_index >= self.sessions.len() && !self.sessions.is_empty() {
-                self.cursor_session_index = self.sessions.len() - 1;
-            }
-            if self.active_session_index >= self.sessions.len() && !self.sessions.is_empty() {
-                self.active_session_index = self.sessions.len() - 1;
-            }
-        }
-    }
-
-    fn refresh_diagnostics(&mut self) {
-        if let Ok(checks) = run_diagnostics(self.library_path.clone()) {
-            self.diagnostics = Some(checks);
-        }
-    }
-
-    fn get_selected_session_target(&self) -> SessionTarget {
-        if self.sessions.is_empty() || self.active_session_index >= self.sessions.len() {
-            SessionTarget::Latest {
-                library: self.library_path.clone(),
-            }
-        } else {
-            SessionTarget::Path(self.sessions[self.active_session_index].path.clone())
-        }
-    }
-
-    fn set_status(&mut self, msg: impl Into<String>) {
-        self.status_message = Some((msg.into(), SystemTime::now()));
-    }
-}
-
-pub fn run_tui(library: Option<PathBuf>) -> anyhow::Result<()> {
-    tracing::info!("Launching interactive TUI...");
-    let mut state = TuiState::new(library).map_err(|e| anyhow::anyhow!(e))?;
-    state.refresh_sessions();
-    state.refresh_diagnostics();
-
-    enable_raw_mode()?;
-    let mut stdout = io::stdout();
-    stdout.execute(EnterAlternateScreen)?;
-    let backend = CrosstermBackend::new(stdout);
-    let mut terminal = Terminal::new(backend)?;
-
-    let res = tui_loop(&mut terminal, state);
-
-    disable_raw_mode()?;
-    io::stdout().execute(LeaveAlternateScreen)?;
-    terminal.show_cursor()?;
-
-    tracing::info!("TUI loop exited. Restored terminal raw modes.");
-    res
-}
-
-fn tui_loop(
-    terminal: &mut Terminal<CrosstermBackend<Stdout>>,
-    mut state: TuiState,
-) -> anyhow::Result<()> {
-    let (tx, rx) = mpsc::channel::<TuiMessage>();
-
-    loop {
-
-        // Render TUI
-        terminal.draw(|f| {
-            draw_ui(f, &state);
-        })?;
-
-        // Check for terminal keyboard input events
-        if event::poll(Duration::from_millis(50))? {
-            if let Event::Key(key) = event::read()? {
-                if key.kind == KeyEventKind::Press {
-                    if state.show_welcome {
-                        match key.code {
-                            KeyCode::Char('q') | KeyCode::Char('Q') => {
-                                break;
-                            }
-                            KeyCode::Char('t') | KeyCode::Char('T') => {
-                                let next_theme = match state.config.theme {
-                                    Theme::Modern => Theme::Classic,
-                                    Theme::Classic => Theme::Modern,
-                                };
-                                state.config.theme = next_theme;
-                                if let Err(e) = state.config.save() {
-                                    state.set_status(format!("Failed to save config: {}", e));
-                                } else {
-                                    state.set_status(format!("Theme toggled to {:?}", next_theme));
-                                }
-                            }
-                            _ => {
-                                state.show_welcome = false;
-                            }
-                        }
-                        continue;
-                    }
-                    if let Some(ref mut input_str) = state.change_library_input {
-                        match key.code {
-                            KeyCode::Enter => {
-                                let new_path_str = input_str.trim().to_string();
-                                if !new_path_str.is_empty() {
-                                    let new_path = PathBuf::from(new_path_str);
-                                    state.library_path = Some(new_path.clone());
-                                    state.resolved_library_path = new_path;
-                                    state.refresh_sessions();
-                                    state.refresh_diagnostics();
-                                    state.set_status("Library path updated successfully");
-                                }
-                                state.change_library_input = None;
-                            }
-                            KeyCode::Esc => {
-                                state.change_library_input = None;
-                            }
-                            KeyCode::Backspace => {
-                                input_str.pop();
-                            }
-                            KeyCode::Char(c) => {
-                                if input_str.len() < 120 {
-                                    input_str.push(c);
-                                }
-                            }
-                            _ => {}
-                        }
-                    } else {
-                        match key.code {
-                            KeyCode::Char('q') | KeyCode::Char('Q') => {
-                                // If capturing, stop it first
-                                if let CaptureState::Capturing { ref stop_handle, .. } = state.capture_state {
-                                    stop_handle.store(true, Ordering::SeqCst);
-                                }
-                                break;
-                            }
-                            KeyCode::Char('l') | KeyCode::Char('L') => {
-                                if let CaptureState::Idle | CaptureState::Error(_) = state.capture_state {
-                                    state.change_library_input = Some(state.resolved_library_path.to_string_lossy().into_owned());
-                                    state.confirm_clean_index = None;
-                                } else {
-                                    state.set_status("Cannot change library while capturing");
-                                }
-                            }
-                            KeyCode::Tab => {
-                                state.active_tab = state.active_tab.next();
-                                if state.active_tab == ActiveTab::Sessions {
-                                    state.refresh_sessions();
-                                } else if state.active_tab == ActiveTab::Diagnostics {
-                                    state.refresh_diagnostics();
-                                }
-                                state.confirm_clean_index = None;
-                            }
-                            KeyCode::Left => {
-                                state.active_tab = state.active_tab.prev();
-                                if state.active_tab == ActiveTab::Sessions {
-                                    state.refresh_sessions();
-                                } else if state.active_tab == ActiveTab::Diagnostics {
-                                    state.refresh_diagnostics();
-                                }
-                                state.confirm_clean_index = None;
-                            }
-                            KeyCode::Right => {
-                                state.active_tab = state.active_tab.next();
-                                if state.active_tab == ActiveTab::Sessions {
-                                    state.refresh_sessions();
-                                } else if state.active_tab == ActiveTab::Diagnostics {
-                                    state.refresh_diagnostics();
-                                }
-                                state.confirm_clean_index = None;
-                            }
-                            KeyCode::Char('1') => {
-                                state.active_tab = ActiveTab::Capture;
-                                state.confirm_clean_index = None;
-                            }
-                            KeyCode::Char('2') => {
-                                state.active_tab = ActiveTab::Render;
-                                state.confirm_clean_index = None;
-                            }
-                            KeyCode::Char('3') => {
-                                state.active_tab = ActiveTab::Sessions;
-                                state.refresh_sessions();
-                                state.confirm_clean_index = None;
-                            }
-                            KeyCode::Char('4') => {
-                                state.active_tab = ActiveTab::Diagnostics;
-                                state.refresh_diagnostics();
-                                state.confirm_clean_index = None;
-                            }
-                            _ => handle_tab_input(&mut state, &key.code, &tx),
-                        }
-                    }
-                }
-            }
-        }
-
-        // Process any messages from background worker threads
-        while let Ok(msg) = rx.try_recv() {
-            match msg {
-                TuiMessage::CaptureStarted(stop_handle) => {
-                    tracing::info!("TUI background capture started successfully.");
-                    state.capture_state = CaptureState::Capturing {
-                        frames_collected: 0,
-                        stop_handle,
-                    };
-                    state.set_status("Capture session started");
-                }
-                TuiMessage::FrameCaptured(index) => {
-                    tracing::debug!("TUI received frame captured event (index: {}).", index);
-                    if let CaptureState::Capturing { ref mut frames_collected, .. } = state.capture_state {
-                        *frames_collected = index + 1;
-                    }
-                }
-                TuiMessage::CaptureFinished(count) => {
-                    tracing::info!("TUI background capture stopped. Saved {} frames.", count);
-                    state.capture_state = CaptureState::Idle;
-                    state.set_status(format!("Capture stopped. Saved {} frames.", count));
-                    state.refresh_sessions();
-                }
-                TuiMessage::CaptureError(err) => {
-                    tracing::error!("TUI background capture thread error: {}", err);
-                    state.capture_state = CaptureState::Error(err.clone());
-                    state.set_status(format!("Capture error: {}", err));
-                }
-                TuiMessage::RenderProgress(msg) => {
-                    tracing::debug!("TUI render progress: {}", msg);
-                    state.render_state = RenderState::Rendering(msg);
-                }
-                TuiMessage::RenderFinished(msg) => {
-                    tracing::info!("TUI render thread finished: {}", msg);
-                    state.render_state = RenderState::Success(msg.clone());
-                    state.set_status("Render completed successfully");
-                    state.refresh_sessions();
-                }
-                TuiMessage::RenderError(err) => {
-                    tracing::error!("TUI render thread error: {}", err);
-                    state.render_state = RenderState::Error(err.clone());
-                    state.set_status(format!("Render failed: {}", err));
-                }
-            }
-        }
-    }
-
-    Ok(())
-}
-
-fn handle_tab_input(state: &mut TuiState, key: &KeyCode, tx: &Sender<TuiMessage>) {
-    match state.active_tab {
-        ActiveTab::Capture => match key {
-            KeyCode::Char(' ') => match state.capture_state {
-                CaptureState::Idle | CaptureState::Error(_) => {
-                    let force = state.capture_mode == CaptureMode::Append;
-                    start_capture_thread(state, tx.clone(), force);
-                }
-                CaptureState::Capturing { ref stop_handle, .. } => {
-                    stop_handle.store(true, Ordering::SeqCst);
-                }
-                CaptureState::Starting => {}
-            },
-            KeyCode::Char('d') | KeyCode::Char('D') => {
-                if let CaptureState::Idle = state.capture_state {
-                    state.capture_display = match state.capture_display {
-                        DisplayTarget::All => DisplayTarget::Primary,
-                        DisplayTarget::Primary => DisplayTarget::All,
-                    };
-                }
-            }
-            KeyCode::Char('a') | KeyCode::Char('A') => {
-                if let CaptureState::Idle = state.capture_state {
-                    state.capture_mode = match state.capture_mode {
-                        CaptureMode::NewSession => {
-                            if state.sessions.is_empty() {
-                                state.set_status("No existing sessions found to append to!");
-                                CaptureMode::NewSession
-                            } else {
-                                CaptureMode::Append
-                            }
-                        }
-                        CaptureMode::Append => CaptureMode::NewSession,
-                    };
-                }
-            }
-            KeyCode::Up => {
-                if let CaptureState::Idle = state.capture_state {
-                    let secs = state.capture_interval.as_secs();
-                    state.capture_interval = Duration::from_secs(secs.saturating_add(1));
-                }
-            }
-            KeyCode::Down => {
-                if let CaptureState::Idle = state.capture_state {
-                    let secs = state.capture_interval.as_secs();
-                    if secs > 1 {
-                        state.capture_interval = Duration::from_secs(secs.saturating_sub(1));
-                    }
-                }
-            }
-            _ => {}
-        },
-        ActiveTab::Render => match key {
-            KeyCode::Char('r') | KeyCode::Char('R') | KeyCode::Enter => {
-                if let RenderState::Idle | RenderState::Success(_) | RenderState::Error(_) =
-                    state.render_state
-                {
-                    start_render_thread(state, tx.clone());
-                }
-            }
-            KeyCode::Up => {
-                state.render_fps = state.render_fps.saturating_add(1);
-            }
-            KeyCode::Down => {
-                if state.render_fps > 1 {
-                    state.render_fps = state.render_fps.saturating_sub(1);
-                }
-            }
-            _ => {}
-        },
-        ActiveTab::Sessions => {
-            if let Some(index) = state.confirm_clean_index {
-                let session = &state.sessions[index];
-                match key {
-                    KeyCode::Char('f') | KeyCode::Char('F') => {
-                        let clean_options = CleanOptions {
-                            target: SessionTarget::Path(session.path.clone()),
-                            frames: true,
-                            videos: false,
-                        };
-                        execute_tui_clean(state, clean_options, false);
-                        state.confirm_clean_index = None;
-                        state.refresh_sessions();
-                    }
-                    KeyCode::Char('v') | KeyCode::Char('V') => {
-                        let clean_options = CleanOptions {
-                            target: SessionTarget::Path(session.path.clone()),
-                            frames: false,
-                            videos: true,
-                        };
-                        execute_tui_clean(state, clean_options, false);
-                        state.confirm_clean_index = None;
-                        state.refresh_sessions();
-                    }
-                    KeyCode::Char('a') | KeyCode::Char('A') => {
-                        let clean_options = CleanOptions {
-                            target: SessionTarget::Path(session.path.clone()),
-                            frames: true,
-                            videos: true,
-                        };
-                        execute_tui_clean(state, clean_options, false);
-                        state.confirm_clean_index = None;
-                        state.refresh_sessions();
-                    }
-                    KeyCode::Char('d') | KeyCode::Char('D') => {
-                        let clean_options = CleanOptions {
-                            target: SessionTarget::Path(session.path.clone()),
-                            frames: true,
-                            videos: true,
-                        };
-                        execute_tui_clean(state, clean_options, true);
-                        state.confirm_clean_index = None;
-                        state.refresh_sessions();
-                    }
-                    _ => {
-                        state.confirm_clean_index = None;
-                    }
-                }
-            } else {
-                match key {
-                    KeyCode::Up => {
-                        if !state.sessions.is_empty() && state.cursor_session_index > 0 {
-                            state.cursor_session_index -= 1;
-                        }
-                    }
-                    KeyCode::Down => {
-                        if !state.sessions.is_empty()
-                            && state.cursor_session_index + 1 < state.sessions.len()
-                        {
-                            state.cursor_session_index += 1;
-                        }
-                    }
-                    KeyCode::Enter | KeyCode::Char(' ') => {
-                        if !state.sessions.is_empty() && state.cursor_session_index < state.sessions.len() {
-                            state.active_session_index = state.cursor_session_index;
-                            let session_name = state.sessions[state.active_session_index].name.clone();
-                            state.set_status(format!("Selected session: {}", session_name));
-                        }
-                    }
-                    KeyCode::Char('o') | KeyCode::Char('O') => {
-                        if !state.sessions.is_empty() {
-                            let target = SessionTarget::Path(state.sessions[state.cursor_session_index].path.clone());
-                            match open_session(target) {
-                                Ok(_) => state.set_status("Opened session in file manager"),
-                                Err(e) => state.set_status(format!("Failed to open: {}", e)),
-                            }
-                        }
-                    }
-                    KeyCode::Char('c') | KeyCode::Char('C') => {
-                        if !state.sessions.is_empty() {
-                            state.confirm_clean_index = Some(state.cursor_session_index);
-                        }
-                    }
-                    KeyCode::Char('a') | KeyCode::Char('A') => {
-                        if !state.sessions.is_empty() {
-                            state.active_session_index = state.cursor_session_index;
-                            state.capture_mode = CaptureMode::Append;
-                            state.active_tab = ActiveTab::Capture;
-                            let session_name = state.sessions[state.active_session_index].name.clone();
-                            state.set_status(format!("Switched to Capture (Append mode for session: {})", session_name));
-                        }
-                    }
-                    KeyCode::Char('u') | KeyCode::Char('U') => {
-                        state.refresh_sessions();
-                        state.set_status("Refreshed sessions list");
-                    }
-                    _ => {}
-                }
-            }
-        }
-        ActiveTab::Diagnostics => match key {
-            KeyCode::Char('d') | KeyCode::Char('D') | KeyCode::Char('u') | KeyCode::Char('U') => {
-                state.refresh_diagnostics();
-                state.set_status("Diagnostics rerun completed");
-            }
-            _ => {}
-        },
-    }
-}
-
-fn start_capture_thread(state: &mut TuiState, tx: Sender<TuiMessage>, force: bool) {
-    tracing::info!(
-        "TUI: Starting capture thread background worker (mode: {:?}, interval: {}s, display: {:?}, force: {})",
-        state.capture_mode,
-        state.capture_interval.as_secs(),
-        state.capture_display,
-        force
-    );
-    state.capture_state = CaptureState::Starting;
-    let library_path = state.library_path.clone();
-    let interval = state.capture_interval;
-    let display = state.capture_display;
-
-    // Resolve append mode target session path
-    let session_path = if state.capture_mode == CaptureMode::Append && !state.sessions.is_empty() {
-        Some(state.sessions[state.active_session_index].path.clone())
-    } else {
-        None
-    };
-    let append = state.capture_mode == CaptureMode::Append && session_path.is_some();
-
-    thread::spawn(move || {
-        let open_options = SessionOpenOptions {
-            library: library_path,
-            session: session_path,
-            append,
-            interval: Some(interval),
-            display,
-            force, // Use the force parameter passed
-        };
-
-        let mut session = match Session::open(open_options) {
-            Ok(s) => s,
-            Err(e) => {
-                let _ = tx.send(TuiMessage::CaptureError(e.to_string()));
-                return;
-            }
-        };
-
-        let capture_loop = CaptureLoop::new(session.capture_interval(), display);
-        let stop_handle = capture_loop.stop_handle();
-
-        if tx.send(TuiMessage::CaptureStarted(stop_handle)).is_err() {
-            return;
-        }
-
-        let backend = XcapBackend;
-        match capture_loop.run_with_progress(&mut session, &backend, |frame_index| {
-            let _ = tx.send(TuiMessage::FrameCaptured(frame_index));
-            Ok(())
-        }) {
-            Ok(count) => {
-                let _ = tx.send(TuiMessage::CaptureFinished(count));
-            }
-            Err(e) => {
-                let _ = tx.send(TuiMessage::CaptureError(e.to_string()));
-            }
-        }
-    });
-}
-
-fn start_render_thread(state: &mut TuiState, tx: Sender<TuiMessage>) {
-    tracing::info!("TUI: Starting render thread background worker...");
-    state.render_state = RenderState::Rendering("Initializing render plan...".to_string());
-    let target = state.get_selected_session_target();
-    let render_target = match target {
-        SessionTarget::Latest { library } => RenderTarget::Latest { library },
-        SessionTarget::Path(path) => RenderTarget::Path(path),
-    };
-    let fps = state.render_fps;
-
-    thread::spawn(move || {
-        let options = RenderOptions {
-            target: render_target,
-            fps,
-            output: None,
-            overwrite: true,
-            verbose: false,
-            exclude: None,
-        };
-
-        let _ = tx.send(TuiMessage::RenderProgress("Running ffmpeg...".to_string()));
-        match render(options) {
-            Ok(result) => {
-                let _ = tx.send(TuiMessage::RenderFinished(format!(
-                    "Rendered {} frames successfully to {}",
-                    result.frame_count,
-                    result.output_path.display()
-                )));
-            }
-            Err(e) => {
-                let _ = tx.send(TuiMessage::RenderError(e.to_string()));
-            }
-        }
-    });
-}
-
-fn execute_tui_clean(state: &mut TuiState, clean_options: CleanOptions, dry_run: bool) {
-    match create_clean_plan(clean_options) {
-        Ok(plan) => {
-            if dry_run {
-                state.set_status(format!(
-                    "Dry run: would delete {} frame(s) and {} video(s)",
-                    plan.frames.len(),
-                    plan.videos.len()
-                ));
-            } else {
-                match execute_clean_plan(&plan) {
-                    Ok(result) => {
-                        state.set_status(format!(
-                            "Cleaned {} frame(s) and {} video(s)",
-                            result.frames_deleted, result.videos_deleted
-                        ));
-                    }
-                    Err(e) => {
-                        state.set_status(format!("Clean execution failed: {}", e));
-                    }
-                }
-            }
-        }
-        Err(e) => {
-            state.set_status(format!("Clean plan failed: {}", e));
-        }
-    }
-}
-
-fn draw_ui(f: &mut ratatui::Frame, state: &TuiState) {
+pub fn draw_ui(f: &mut ratatui::Frame, state: &TuiState) {
     let size = f.area();
 
-    // Screen size warning guard (Option 1)
+    // Screen size warning guard
     if size.width < 80 || size.height < 20 {
         let msg = format!(
             "\n\n  Terminal window size is too small!\n\n  \
